[#160192048] Refactor logic from HOC to render props
