--- src/_utils/checkPropertyExists.js
@@ -1,1 +1,0 @@
-export default (object, property) => Object.prototype.hasOwnProperty.call(object, property);
